+++ GR_Datasets.lua
@@ -10,7 +10,7 @@ function ds:Init()
     self.tblClassesByName = nil
 
     self.tblBadZones = ds:invalidZones()
-    self.tblBadZonesByName = nil
+    self.tblBadZonesByName = ds:invalidZonesByName()
 
     self.tblWhispers = {}
     self.tblConnected = ds:GetConnectedRealms()
@@ -39,6 +39,7 @@ function ds:WhatsNew()
             (Found in settings, Invite Settings).
             |CFF55D0FFNote: This does not overwrite any existing
             keybindings in WoW.|r
+        * Players in instances will not be invited or messaged.
 
     |CFFFFFF00v2.1.33 Notes|r
         * Brought back adding black list players to the icon bar.
@@ -216,59 +217,73 @@ end
 function ds:invalidZones()
     local tbl = {
         --battlegrounds
-        [30] = "Alterac Valley",
-        [489] = "Warsong Gulch",
-        [529] = "Arathi Basin (Classic)",
-        [566] = "Eye of the Storm",
-        [607] = "Strand of the Ancients",
-        [628] = "Isle of Conquest",
-        [726] = "Twin Peaks",
-        [727] = "Silvershard Mines",
-        [761] = "The Battle for Gilneas",
-        [968] = "Eye of the Storm (Rated)",
-        [998] = "Temple of Kotmogu",
-        [1105] = "Deepwind Gorge",
-        [1681] = "Arathi Basin (Winter)",
-        [1803] = "Seething Shore",
-        [2107] = "Arathi Basin",
-        [2177] = "Arathi Basin Comp Stomp",
+        [30] = { name = "Alterac Valley", reason = 'Battlegrounds' },
+        [489] = { name = "Warsong Gulch", reason = 'Battlegrounds' },
+        [529] = { name = "Arathi Basin (Classic)", reason = 'Battlegrounds' },
+        [566] = { name = "Eye of the Storm", reason = 'Battlegrounds' },
+        [607] = { name = "Strand of the Ancients", reason = 'Battlegrounds' },
+        [628] = { name = "Isle of Conquest", reason = 'Battlegrounds' },
+        [726] = { name = "Twin Peaks", reason = 'Battlegrounds' },
+        [727] = { name = "Silvershard Mines", reason = 'Battlegrounds' },
+        [761] = { name = "The Battle for Gilneas", reason = 'Battlegrounds' },
+        [968] = { name = "Eye of the Storm (Rated)", reason = 'Battlegrounds' },
+        [998] = { name = "Temple of Kotmogu", reason = 'Battlegrounds' },
+        [1105] = { name = "Deepwind Gorge", reason = 'Battlegrounds' },
+        [1681] = { name = "Arathi Basin (Winter)", reason = 'Battlegrounds' },
+        [1803] = { name = "Seething Shore", reason = 'Battlegrounds' },
+        [2107] = { name = "Arathi Basin", reason = 'Battlegrounds' },
+        [2177] = { name = "Arathi Basin Comp Stomp", reason = 'Battlegrounds' },
         --arenas
-        [572] = "Ruins of Lordaeron",
-        [617] = "Dalaran Arena",
-        [618] = "The Ring of Valor",
-        [980] = "Tol'Viron Arena",
-        [1134] = "Tiger's Peak",
-        [1505] = "Nagrand Arena",
-        [1672] = "Blade's Edge Arena",
-        [2167] = "The Robodrome",
+        [572] = { name = "Ruins of Lordaeron", reason = 'Arena' },
+        [617] = { name = "Dalaran Arena", reason = 'Arena' },
+        [618] = { name = "The Ring of Valor", reason = 'Arena' },
+        [980] = { name = "Tol'Viron Arena", reason = 'Arena' },
+        [1134] = { name = "Tiger's Peak", reason = 'Arena' },
+        [1505] = { name = "Nagrand Arena", reason = 'Arena' },
+        [1672] = { name = "Blade's Edge Arena", reason = 'Arena' },
+        [2167] = { name = "The Robodrome", reason = 'Arena' },
 
             --@version-retail@
         --raids
-        [2522] = "Vault of the Incarnates",
-        [2569] = "Aberrus, the Shadowed Crucible",
+        [11111] = { name = "Amirdrassil, the Dream's Hope", reason = 'Season 3 Raid' },
+        [2522] = { name = "Vault of the Incarnates", reason = 'Season 2 Raid' },
+        --[2569] = { name = "Aberrus, the Shadowed Crucible", reason = 'Season 1 Raid' },
         --dungeons
-        [2451] = "Uldaman: Legacy of Tyr",
-        [2515] = "The Azure Vault",
-        [2516] = "The Nokhud Offensive",
-        [2519] = "Neltharus",
-        [2520] = "Brackenhide Hollow",
-        [2521] = "Ruby Life Pools",
-        [2526] = "Algeth'ar Academy",
-        [2527] = "Halls of Infusion",
+        --[2451] = { name = "Uldaman: Legacy of Tyr", reason = 'DF Dungeon' },
+        --[2515] = { name = "The Azure Vault", reason = 'DF Dungeon' },
+        --[2516] = { name = "The Nokhud Offensive", reason = 'DF Dungeon' },
+        --[2519] = { name = "Neltharus", reason = 'DF Dungeon' },
+        --[2520] = { name = "Brackenhide Hollow", reason = 'DF Dungeon' },
+        --[2521] = { name = "Ruby Life Pools", reason = 'DF Dungeon' },
+        --[2526] = { name = "Algeth'ar Academy", reason = 'DF Dungeon' },
+        --[2527] = { name = "Halls of Infusion", reason = 'DF Dungeon' },
             --M+ rotating
-        [1754] = "Freehold",
-        [1841] = "The Underrot",
-        [1458] = "Neltharion's Lair",
-        [657] = "The Vortex Pinnacle",
+        [1754] = { name = "Freehold", reason = 'Season 2 Dungeon' },
+        [1841] = { name = "The Underrot", reason = 'Season 2 Dungeon' },
+        [1458] = { name = "Neltharion's Lair", reason = 'Season 2 Dungeon' },
+        [657] = { name = "The Vortex Pinnacle", reason = 'Season 2 Dungeon' },
+        [1466] = { name = "Darkheart Thicket", reason = 'Season 3 Dungeon' },
+        [1501] = { name = 'Black Rook Hold', reason = 'Season 3 Dungeon' },
+        [1862] = { name = 'Waycrest Manor', reason = 'Season 3 Dungeon' },
+        [1763] = { name = "Atal'Dazar", reason = 'Season 3 Dungeon' },
+        [1279] = { name = 'The Everbloom', reason = 'Season 3 Dungeon' },
+        [643] = { name = 'Throne of Tides', reason = 'Season 3 Dungeon' },
+        [22222] = { name = "Dawn of the Infinite: Galakrond's Fall", reason = 'Season 3 Dungeon' },
+        [33333] = { name = "Dawn of the Infinite Murozond's Rise", reason = 'Season 3 Dungeon' },
     }
 
     return tbl
 end
 function ds:invalidZonesByName()
-    self.tblBadZonesByName = {}
-    for k, r in pairs(self.tblBadZones) do
-        self.tblBadZonesByName[r] = k
+    local tbl = ds:invalidZones()
+    local tblOut = {}
+    for k, r in pairs(tbl) do
+        local name = r.name:gsub("%b()", ""):trim()
+        if not tblOut[strlower(name)] then
+            tblOut[strlower(name)] = { id = k, name = name, reason = r.reason } end
     end
+
+    return tblOut
 end
 function ds:WhisperMessages(performCheck)
     ns.db.messages = ns.db.messages or {}